refactor: dao를 interface와 구현체로 수정함
