+++ src/main/java/cart/dao/CartProductDao.java
@@ -1,0 +1,9 @@
+package cart.dao;
+
+import cart.dto.CartProductResponse;
+
+import java.util.List;
+
+public interface CartProductDao {
+    List<CartProductResponse> selectById(String email);
+}
