+++ src/main/java/cart/service/CartService.java
@@ -1,6 +1,7 @@
 package cart.service;
 
-import cart.dao.CartDao;
+import cart.dao.h2Implement.CartH2Dao;
+import cart.dao.h2Implement.CartProductH2Dao;
 import cart.dto.CartProductResponse;
 import cart.entity.Cart;
 import org.springframework.stereotype.Service;
@@ -11,24 +12,27 @@ import java.util.List;
 @Service
 @Transactional
 public class CartService {
-    private final CartDao cartDao;
+    private final CartH2Dao cartH2Dao;
 
-    public CartService(CartDao cartDao) {
-        this.cartDao = cartDao;
+    private final CartProductH2Dao cartProductH2Dao;
+
+    public CartService(CartH2Dao cartH2Dao, CartProductH2Dao cartProductH2Dao) {
+        this.cartH2Dao = cartH2Dao;
+        this.cartProductH2Dao = cartProductH2Dao;
     }
 
     @Transactional(readOnly = true)
     public List<CartProductResponse> getCartsByEmail(String email) {
-        return cartDao.selectById(email);
+        return cartProductH2Dao.selectById(email);
     }
 
     public int addCart(int productId, String email) {
         Cart cart = new Cart(productId, email);
-        return cartDao.insert(cart);
+        return cartH2Dao.insert(cart);
     }
 
     public void deleteCart(int cartId, String email) {
-        int deleteResult = cartDao.delete(cartId, email);
+        int deleteResult = cartH2Dao.delete(cartId, email);
         if (deleteResult == 0) {
             throw new IllegalStateException("존재하지 않는 장바구니 입니다.");
         }
