+++ src/main/java/cart/service/ProductService.java
@@ -1,6 +1,6 @@
 package cart.service;
 
-import cart.dao.ProductDao;
+import cart.dao.h2Implement.ProductH2Dao;
 import cart.dto.ProductRequest;
 import cart.dto.ProductResponse;
 import cart.entity.Product;
@@ -13,21 +13,21 @@ import java.util.stream.Collectors;
 @Service
 @Transactional
 public class ProductService {
-    private final ProductDao productDao;
+    private final ProductH2Dao productH2Dao;
 
-    public ProductService(ProductDao productDao) {
-        this.productDao = productDao;
+    public ProductService(ProductH2Dao productH2Dao) {
+        this.productH2Dao = productH2Dao;
     }
 
     public int addProduct(ProductRequest productRequest) {
         Product product = new Product(productRequest.getName(), productRequest.getPrice(),
                 productRequest.getImage());
-        return productDao.insert(product);
+        return productH2Dao.insert(product);
     }
 
     @Transactional(readOnly = true)
     public List<ProductResponse> selectAllProducts() {
-        final List<Product> products = productDao.selectAll();
+        final List<Product> products = productH2Dao.selectAll();
         return products.stream()
                 .map(ProductResponse::from)
                 .collect(Collectors.toUnmodifiableList());
@@ -36,14 +36,14 @@ public class ProductService {
     public void updateProduct(ProductRequest productRequest, int productId) {
         Product product = new Product(productId, productRequest.getName(),
                 productRequest.getPrice(), productRequest.getImage());
-        int updatedResult = productDao.update(product);
+        int updatedResult = productH2Dao.update(product);
         if (updatedResult == 0) {
             throw new IllegalStateException("존재하지 않는 상품입니다.");
         }
     }
 
     public void deleteProduct(int productId) {
-        int deletedResult = productDao.delete(productId);
+        int deletedResult = productH2Dao.delete(productId);
         if (deletedResult == 0) {
             throw new IllegalStateException("존재하지 않는 상품입니다.");
         }
